Rename channel source files for casing scheme
