+++ src/webR/chan/channel-shared.ts
